requirements.txt
requests
moviepy
opencv-python
numpy
Pillow
selenium
pytubefix
librosa
scipy
matplotlib
python-dotenv
